+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Category.java
@@ -14,4 +14,11 @@ import javax.persistence.Table;
 public class Category extends BaseEntity {
     private String name;
     private String description;
+
+    @Builder
+    public Category(Long id, String name, String description) {
+        super(id);
+        this.name = name;
+        this.description = description;
+    }
 }