+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Company.java
@@ -9,4 +9,11 @@ import lombok.*;
 public class Company extends BaseEntity {
     private String title;
     private String description;
+
+    @Builder
+    public Company(Long id, String title, String description) {
+        super(id);
+        this.title = title;
+        this.description = description;
+    }
 }